+++ src/cider/nrepl.clj
@@ -342,7 +342,7 @@ if applicable, and re-render the updated value."
                           "view-mode" "Mode of viewing the value - either `:normal` or `:object`"}
                :returns inspector-returns}
               "inspect-toggle-view-mode"
-              {:doc "Toggles the viewing mode of the inspector. This influences the way how inspector is rendering the current value. `:normal` is the default. When view mode is `:object`, any value will be rendered as a Java object (fields shown as is). View mode is automatically reset back to normal when navigating to child values."
+              {:doc "Toggles the viewing mode of the inspector. This influences the way how inspector is rendering the current value. `:normal` is the default. When view mode is `:table`, the value will be rendered as a table (only supported for sequences of maps or tuples). When view mode is `:object`, any value will be rendered as a Java object (fields shown as is). View mode is automatically reset back to normal when navigating to child values."
                :requires {"session" "The current session"}
                :returns inspector-returns}
               "inspect-display-analytics"